requirements.txt
pandas
sentence-transformers
chromadb
requests
pyyaml
torch
numpy
python-dotenv
openai>=1.0.0